Update Volume.tokens() to return set
This also fixes an earlier bug. A set is a fast way to get unique
tokens, and sensible for this particular method, where intersections
between vocabularies are desired. Converting to a list or Series
internally would be a waste of processing if a user will just convert
back; rather, then can cast to list or Series easily if needed.

Since this is a format change, it seemed right to iterate the version
number, but I don't really know what I'm doing with numbering anymore.